Add files via upload
Penal Law Section 135.60
Coercion in the third degree
A person is guilty of coercion in the third degree when he or she compels or induces a person to engage in conduct which the latter has a legal right to abstain from engaging in, or to abstain from engaging in conduct in which he or she has a legal right to engage, or compels or induces a person to join a group, organization or criminal enterprise which such latter person has a right to abstain from joining, or compels or induces a person to produce, disseminate, or otherwise display an image or images depicting nudity of such person or depicting such person engaged in sexual conduct as defined in subdivisions two and three of section 235.20 of this chapter, by means of instilling in him or her a fear that, if the demand is not complied with, the actor or another will:
1.Cause physical injury to a person; or
2.Cause damage to property; or
3.Engage in other conduct constituting a crime; or
4.Accuse some person of a crime or cause criminal charges to be instituted against him or her; or
5.Expose a secret or publicize an asserted fact, whether true or false, tending to subject some person to hatred, contempt or ridicule; or
6.Cause a strike, boycott or other collective labor group action injurious to some person’s business; except that such a threat shall not be deemed coercive when the act or omission compelled is for the benefit of the group in whose interest the actor purports to act; or
7.Testify or provide information or withhold testimony or information with respect to another’s legal claim or defense; or
8.Use or abuse his or her position as a public servant by performing some act within or related to his or her official duties, or by failing or refusing to perform an official duty, in such manner as to affect some person adversely; or
9.Perform any other act which would not in itself materially benefit the actor but which is calculated to harm another person materially with respect to his or her health, safety, business, calling, career, financial condition, reputation or personal relationships.
10.Report his or her immigration status or suspected immigration status. Coercion in the third degree is a class A misdemeanor.
Source: Section 135.60 — Coercion in the third degree, https://www.­nysenate.­gov/legislation/laws/PEN/135.­60 (updated Dec. 24, 2021; accessed Aug. 31, 2024).